Collider offset (#109)
* cleanup: some cast warning and remove NestedObject

d3d size_t to uint

* collider: add offset

* water: test collider offset

* object: addcomponent returns generated component

* test: multiple colliders

* pch: cleanup

+++ Engine/egApplication.hpp
@@ -37,7 +37,7 @@ namespace Engine::Manager
 
         float GetDeltaTime() const
         {
-            return m_timer->GetElapsedSeconds();
+            return static_cast<float>(m_timer->GetElapsedSeconds());
         }
 
         uint32_t GetFPS() const
